Shard code.

+++ java/modules/basic/src/main/java/io/v6d/modules/basic/arrow/BooleanArray.java
@@ -36,7 +36,7 @@ public class BooleanArray extends Array {
         super(meta);
         this.array = new BitVector("", Arrow.default_allocator);
         this.array.loadFieldBuffers(
-                new ArrowFieldNode(length, 0), Arrays.asList(null, buffer.getBuffer()));
+                new ArrowFieldNode((int)length, 0), Arrays.asList(null, buffer.getBuffer()));
     }
 
     public double get(int index) {
