+++ java/modules/basic/src/main/java/io/v6d/modules/basic/arrow/Int64ArrayBuilder.java
@@ -32,7 +32,7 @@ public class Int64ArrayBuilder implements ArrayBuilder {
         this.array = new BigIntVector("", Arrow.default_allocator);
         this.buffer = new BufferBuilder(client, this.array.getBufferSizeFor((int) length));
         this.array.loadFieldBuffers(
-                new ArrowFieldNode(length, 0), Arrays.asList(null, buffer.getBuffer()));
+                new ArrowFieldNode((int) length, 0), Arrays.asList(null, buffer.getBuffer()));
     }
 
     @Override
